refactor : SecurityContext를 UserAuthService에서 접근하도록 수정
